Filter for only certs expiring in the next 30 days
When running this, it showed all of the certs, with their "before" and
"after" dates.  This was hard to quickly parse through, and made it hard
to tell if there were any entries that we were worried about.

Adding this logic makes it so that it only shows the dates that are
going to expire within the next 30 days.  The way it does this is by:
- setting a default date for comparison which is 30 days in the future
- getting the SSL cert data as we did before, but filtering it down to
  just the "after" date and then cutting out everything except the date
- parsing the after date into YYYYMMDD format
- comparing the after date to the comparison date, and print information
  if after date is earlier than the comparison date.